requirements.txt
streamlit>=1.32
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
pandas>=2.0
pillow>=10.0
pytesseract>=0.3.10
nltk>=3.8
language-tool-python>=2.7
requests>=2.31
tqdm>=4.66

# Providers (optionnels)
openai>=1.0.0
deepl>=1.16.0

# Dev / tests
pytest>=8.0
